requirements.txt
annotated-types==0.7.0
cffi==1.17.1
donfig==0.8.1.post1
iniconfig==2.1.0
llvmlite==0.44.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-graphblas==2025.2.0
pytz==2025.2
PyYAML==6.0.2
scipy==1.15.3
six==1.17.0
suitesparse-graphblas==9.4.5.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
